fixed compilation errors occuring on my desktop ;-) More of them are needed to fix also the py_sems problem.
git-svn-id: http://svn.berlios.de/svnroot/repos/sems/trunk@230 8eb893ce-cfd4-0310-b710-fb5ebe64c474

+++ apps/registrar_client/SIPRegistrarClient.cpp
@@ -301,7 +301,7 @@ bool SIPRegistration::registerSendTimeout(time_t now_sec) {
 bool SIPRegistration::timeToReregister(time_t now_sec) {
 //   	if (active) 
 //   		DBG("compare %lu with %lu\n",(reg_begin+reg_expires), (unsigned long)now_sec);
-	return ((reg_begin+reg_expires/2) < (unsigned long)now_sec);	
+	return (((unsigned long)reg_begin+reg_expires/2) < (unsigned long)now_sec);	
 }
 
 bool SIPRegistration::registerExpired(time_t now_sec) {
